Solana Smart Contracts
I have added the smart contracts that will be used in the App. These smart contracts are :counter->whitelist and the minter.

The idea is that with the phone camera, you could actually whitelist wallets just scanning QR codes with respective Pubkey. Also an implementation will be sending NFT + Whitelisting.